Game + readme update
Rewrote a bunch of code, cleaning things up and added various features to make it more "game-like".

Added models to make it a bit funnier to play against the AI and to make it more interactable. Added difficulty modes and more stable ability to adjust settings as well as to restart the game.

Also AI now runs asynchronously!